Better incremental parsing
With this patch, when a syntax error is encountered,
the remaining of the current line will be discarded before
we try again to parse a statement. This allows to avoid
having residual syntax errors (if a syntax error is detected
in the middle of the line, chances are the rest of the line
is *not* a valid statement).