Parse attributes using application instead of colons
Smtlib attributes can have an optional argument. Until now this was
parsed as a colon, aka juxtaposition. However the type-checker parses
attributes based on application, so this commit changes the parser to
produce applications instead, which is actually a bit more intuitive
given the semantics of an attribute optional value.