Re-enable support for "and" inside function/predicate
In #129 support was added for mutually recursive function and predicate
definitions using the "and" keyword.

However, the keyword "and" can also be used to denote the boolean
conjunection, but now the conflicts are resolved in such a way that:

```
predicate X = true and true
```

is now a syntax error (upon seeing the [and] token, [predicate X = true]
gets reduced into a [predicate_def], and a recursive [function] or
[predicate] definition is expected).

This patch fixes the issue by marking the appropriate rules
(predicate_def, function_def) as inline, propagating enough context to
the conflict that it can be resolved appropriately.

Note that we are still not able to parse horrors such as:

```
predicate X = true and true and predicate Y = true
```

but that is OK because those were never supported by Alt-Ergo.

(as an aside, I don't think it's even possible by using a %right
associative token for AND in Menhir)