Add support for prelude files
This improves Dolmen's prelude support by allowing to provide preludes as
files, rather than as statements.

The prelude files are parsed separately from the main file, and can use a
different language from the main file. The `logic_file` is reset to the
appropriate file during statement generation to ensure that the typer always
uses the right language.

Since a single pipe is now expected to potentially include statements from
different languages, the `additional_builtins` in the typer is passed the
current language as an additional argument, allowing to define
language-specific additional builtins. Since this is already a breaking change,
`additional_builtins` is also changed from a reference to a `State` key for
consistency with the rest of Dolmen's API.