Move checking of transition systems in a separate file
The new code uses the `get_global_custom` and `set_global_custom` functions to
manage the table of defined systems, and calls the check_no_free_wildcards function.
This commit also removes the system constant that is not necessary anymore.
The code that emits warnings for unused variables is enabled again;
no issues were found.