More information for reserved id, and adjusted errors
This commits adds more information to "reserved" ids in the typechecker,
to distinguish ids reserved for solver use (e.g. those starting with '@'
or '.' in smt2), from ids that may be used in some models (such as
`div0`). This allows to correctly decide when to error out when
a reserved id becomes shadowed.