Support SMT-LIB v2.7 bit-vector operators in psmt2
The new bit-vector operators from v2.7 are not available in SMT-LIB 2.6
scripts, which makes sense since Dolmen is intended to be able to check
adherence to the specification.

On the other hand, there is no standard specification for the psmt2
format, and while it implements polymorphism differently from SMT-LIB
2.7, there is no reason for psmt2 operators to be restricted to those
available in version 2.6 of the standard.

Allow the new bit-vector operators from version 2.7 of the standard when
parsing files with psmt2 syntax.